third_party/blink/web_tests/platform/generic/fast/filesystem/simple-required-arguments-remove-expected.txt

Entry.remove() required arguments test.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

webkitRequestFileSystem(TEMPORARY, 100, successCallback, errorCallback);
Successfully obtained TEMPORARY FileSystem:file__0:Temporary
root = fileSystem.root
PASS root.remove() threw exception TypeError: Failed to execute 'remove' on 'Entry': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
